+++ piscine_C11/ex05/ft_func.h
@@ -6,19 +6,19 @@
 /*   By: imgwang-yeol <imgwang-yeol@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/21 12:36:38 by gim               #+#    #+#             */
-/*   Updated: 2020/07/22 14:56:12 by imgwang-yeo      ###   ########.fr       */
+/*   Updated: 2020/07/22 17:13:15 by imgwang-yeo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_FUNC_H
 # define FT_FUNC_H
 
-long long	add(long long n1, long long n2);
-long long	min(long long n1, long long n2);
-long long	mul(long long n1, long long n2);
-long long	div(long long n1, long long n2);
-long long	mod(long long n1, long long n2);
-long long	ft_atoi(char *str);
-void		ft_putnbr(long long nb);
+int 	add(int n1, int n2);
+int 	min(int n1, int n2);
+int 	mul(int n1, int n2);
+int 	div(int n1, int n2);
+int 	mod(int n1, int n2);
+int 	ft_atoi(char *str);
+void	ft_putnbr(int nb);
 
 #endif
