+++ piscine_C11/ex05/ft_putnbr.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gim <gim@student.42.fr>                    +#+  +:+       +#+        */
+/*   By: imgwang-yeol <imgwang-yeol@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/12 12:55:56 by gim               #+#    #+#             */
-/*   Updated: 2020/07/21 13:37:00 by gim              ###   ########.fr       */
+/*   Updated: 2020/07/22 17:11:43 by imgwang-yeo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 
-void	ft_putnbr(long long nb)
+void	ft_putnbr(int nb)
 {
 	if (nb < 0)
 	{
